+++ src/main/java/net/pwall/log/GradleProxy.java
@@ -1,5 +1,5 @@
 /*
- * @(#) Slf4jProxy.java
+ * @(#) GradleProxy.java
  *
  * log-front  Logging interface
  * Copyright (c) 2020, 2021, 2022, 2024 Peter Wall
@@ -29,53 +29,65 @@ import java.lang.reflect.InvocationTargetException;
 import java.lang.reflect.Method;
 
 /**
- * A class to handle calls to the underlying {@code slf4j} implementation using reflection.
+ * A class to handle calls to the Gradle {@code slf4j} implementation using reflection.
+ * <br>
+ * Gradle has its own set of logging levels, and these are mapped as follows:
+ * <dl>
+ *   <dt><b>ERROR</b></dt>
+ *   <dd>Maps to Gradle <b>ERROR</b></dd>
+ *   <dt><b>WARN</b></dt>
+ *   <dd>Maps to Gradle <b>WARN</b></dd>
+ *   <dt><b>INFO</b></dt>
+ *   <dd>Maps to Gradle <b>LIFECYCLE</b></dd>
+ *   <dt><b>DEBUG</b></dt>
+ *   <dd>Maps to Gradle <b>INFO</b></dd>
+ *   <dt><b>TRACE</b></dt>
+ *   <dd>Maps to Gradle <b>DEBUG</b></dd>
+ * </dl>
  *
  * @author  Peter Wall
  */
-public class Slf4jProxy implements Slf4jProxyInterface {
+public class GradleProxy implements Slf4jProxyInterface {
 
     private final Method getLoggerMethod;
 
-    private final Method isTraceEnabledMethod;
-    private final Method isDebugEnabledMethod;
-    private final Method isInfoEnabledMethod;
-    private final Method isWarnEnabledMethod;
-    private final Method isErrorEnabledMethod;
-    private final Method traceMethod;
-    private final Method debugMethod;
-    private final Method infoMethod;
-    private final Method warnMethod;
-    private final Method errorMethod;
-    private final Method errorThrowableMethod;
+    private final Object debugLevel;
+    private final Object infoLevel;
+    private final Object lifecycleLevel;
+    private final Object warnLevel;
+    private final Object errorLevel;
+    private final Object[] levels;
 
-    private final Method[] dynamicIsEnabledMethods;
-    private final Method[] dynamicLogMethods;
+    private final Method isEnabledMethod;
+    private final Method logMethod;
+    private final Method logThrowableMethod;
 
     /**
      * Construct an {@code Slf4jProxy}.  Only one such object will be required.
      *
-     * @throws ClassNotFoundException   if the {@code slf4j} classes are not found
-     * @throws NoSuchMethodException    if the required methods are not found
+     * @throws ClassNotFoundException       if the {@code slf4j} classes are not found
+     * @throws NoSuchMethodException        if the required methods are not found
+     * @throws InvocationTargetException    if calls to Gradle classes fail
+     * @throws IllegalAccessException       if calls to Gradle are not allowed
      */
-    public Slf4jProxy() throws ClassNotFoundException, NoSuchMethodException {
-        Class<?> loggerFactoryClass = Class.forName("org.slf4j.LoggerFactory");
+    public GradleProxy() throws ClassNotFoundException, NoSuchMethodException, InvocationTargetException,
+            IllegalAccessException {
+        Class<?> loggerFactoryClass = Class.forName("org.gradle.api.logging.Logging");
         getLoggerMethod = loggerFactoryClass.getMethod("getLogger", String.class);
-        Class<?> loggerClass = Class.forName("org.slf4j.Logger");
-        isTraceEnabledMethod = loggerClass.getMethod("isTraceEnabled");
-        isDebugEnabledMethod = loggerClass.getMethod("isDebugEnabled");
-        isInfoEnabledMethod = loggerClass.getMethod("isInfoEnabled");
-        isWarnEnabledMethod = loggerClass.getMethod("isWarnEnabled");
-        isErrorEnabledMethod = loggerClass.getMethod("isErrorEnabled");
-        traceMethod = loggerClass.getMethod("trace", String.class);
-        debugMethod = loggerClass.getMethod("debug", String.class);
-        infoMethod = loggerClass.getMethod("info", String.class);
-        warnMethod = loggerClass.getMethod("warn", String.class);
-        errorMethod = loggerClass.getMethod("error", String.class);
-        errorThrowableMethod = loggerClass.getMethod("error", String.class, Throwable.class);
-        dynamicIsEnabledMethods = new Method[] { isTraceEnabledMethod, isDebugEnabledMethod, isInfoEnabledMethod,
-                isWarnEnabledMethod, isErrorEnabledMethod };
-        dynamicLogMethods = new Method[] { traceMethod, debugMethod, infoMethod, warnMethod, errorMethod };
+
+        Class<?> levelClass = Class.forName("org.gradle.api.logging.LogLevel");
+        Method valueOfMethod = levelClass.getMethod("valueOf", String.class);
+        debugLevel = valueOfMethod.invoke(null, "DEBUG");
+        infoLevel = valueOfMethod.invoke(null, "INFO");
+        lifecycleLevel = valueOfMethod.invoke(null, "LIFECYCLE");
+        warnLevel = valueOfMethod.invoke(null, "WARN");
+        errorLevel = valueOfMethod.invoke(null, "ERROR");
+        levels = new Object[] { debugLevel, infoLevel, lifecycleLevel, warnLevel, errorLevel };
+
+        Class<?> loggerClass = Class.forName("org.gradle.api.logging.Logger");
+        isEnabledMethod = loggerClass.getMethod("isEnabled", levelClass);
+        logMethod = loggerClass.getMethod("log", levelClass, String.class);
+        logThrowableMethod = loggerClass.getMethod("log", levelClass, String.class, Throwable.class);
     }
 
     /**
@@ -94,13 +106,13 @@ public class Slf4jProxy implements Slf4jProxyInterface {
     /**
      * Test whether trace output is enabled for this {@code Logger}.
      *
-     * @param   slf4jLogger     the {@code Logger}
+     * @param   gradleLogger    the {@code Logger}
      * @return                  {@code true} if trace output is enabled
      */
     @Override
-    public boolean isTraceEnabled(Object slf4jLogger) {
+    public boolean isTraceEnabled(Object gradleLogger) {
         try {
-            return (Boolean)isTraceEnabledMethod.invoke(slf4jLogger);
+            return (Boolean)isEnabledMethod.invoke(gradleLogger, debugLevel);
         }
         catch (IllegalAccessException | InvocationTargetException e) {
             throw new Slf4jLoggerException(e);
@@ -110,13 +122,13 @@ public class Slf4jProxy implements Slf4jProxyInterface {
     /**
      * Test whether debug output is enabled for this {@code Logger}.
      *
-     * @param   slf4jLogger     the {@code Logger}
+     * @param   gradleLogger    the {@code Logger}
      * @return                  {@code true} if debug output is enabled
      */
     @Override
-    public boolean isDebugEnabled(Object slf4jLogger) {
+    public boolean isDebugEnabled(Object gradleLogger) {
         try {
-            return (Boolean)isDebugEnabledMethod.invoke(slf4jLogger);
+            return (Boolean)isEnabledMethod.invoke(gradleLogger, infoLevel);
         }
         catch (IllegalAccessException | InvocationTargetException e) {
             throw new Slf4jLoggerException(e);
@@ -126,13 +138,13 @@ public class Slf4jProxy implements Slf4jProxyInterface {
     /**
      * Test whether info output is enabled for this {@code Logger}.
      *
-     * @param   slf4jLogger     the {@code Logger}
+     * @param   gradleLogger    the {@code Logger}
      * @return                  {@code true} if info output is enabled
      */
     @Override
-    public boolean isInfoEnabled(Object slf4jLogger) {
+    public boolean isInfoEnabled(Object gradleLogger) {
         try {
-            return (Boolean)isInfoEnabledMethod.invoke(slf4jLogger);
+            return (Boolean)isEnabledMethod.invoke(gradleLogger, lifecycleLevel);
         }
         catch (IllegalAccessException | InvocationTargetException e) {
             throw new Slf4jLoggerException(e);
@@ -142,13 +154,13 @@ public class Slf4jProxy implements Slf4jProxyInterface {
     /**
      * Test whether warning output is enabled for this {@code Logger}.
      *
-     * @param   slf4jLogger     the {@code Logger}
+     * @param   gradleLogger    the {@code Logger}
      * @return                  {@code true} if warning output is enabled
      */
     @Override
-    public boolean isWarnEnabled(Object slf4jLogger) {
+    public boolean isWarnEnabled(Object gradleLogger) {
         try {
-            return (Boolean)isWarnEnabledMethod.invoke(slf4jLogger);
+            return (Boolean)isEnabledMethod.invoke(gradleLogger, warnLevel);
         }
         catch (IllegalAccessException | InvocationTargetException e) {
             throw new Slf4jLoggerException(e);
@@ -158,13 +170,13 @@ public class Slf4jProxy implements Slf4jProxyInterface {
     /**
      * Test whether error output is enabled for this {@code Logger}.
      *
-     * @param   slf4jLogger     the {@code Logger}
+     * @param   gradleLogger    the {@code Logger}
      * @return                  {@code true} if error output is enabled
      */
     @Override
-    public boolean isErrorEnabled(Object slf4jLogger) {
+    public boolean isErrorEnabled(Object gradleLogger) {
         try {
-            return (Boolean)isErrorEnabledMethod.invoke(slf4jLogger);
+            return (Boolean)isEnabledMethod.invoke(gradleLogger, warnLevel);
         }
         catch (IllegalAccessException | InvocationTargetException e) {
             throw new Slf4jLoggerException(e);
@@ -174,14 +186,14 @@ public class Slf4jProxy implements Slf4jProxyInterface {
     /**
      * Test whether a specified level is enabled for this {@code Logger}.
      *
-     * @param   slf4jLogger     the {@code Logger}
+     * @param   gradleLogger    the {@code Logger}
      * @param   level           the {@code Level}
      * @return                  {@code true} if error output is enabled
      */
     @Override
-    public boolean isEnabled(Object slf4jLogger, Level level) {
+    public boolean isEnabled(Object gradleLogger, Level level) {
         try {
-            return (Boolean)dynamicIsEnabledMethods[level.ordinal()].invoke(slf4jLogger);
+            return (Boolean)isEnabledMethod.invoke(gradleLogger, levels[level.ordinal()]);
         }
         catch (IllegalAccessException | InvocationTargetException e) {
             throw new Slf4jLoggerException(e);
@@ -191,13 +203,13 @@ public class Slf4jProxy implements Slf4jProxyInterface {
     /**
      * Output a trace message.
      *
-     * @param   slf4jLogger     the {@code Logger}
+     * @param   gradleLogger    the {@code Logger}
      * @param   text            the message
      */
     @Override
-    public void trace(Object slf4jLogger, String text) {
+    public void trace(Object gradleLogger, String text) {
         try {
-            traceMethod.invoke(slf4jLogger, text);
+            logMethod.invoke(gradleLogger, debugLevel, text);
         }
         catch (IllegalAccessException | InvocationTargetException e) {
             throw new Slf4jLoggerException(e);
@@ -207,13 +219,13 @@ public class Slf4jProxy implements Slf4jProxyInterface {
     /**
      * Output a debug message.
      *
-     * @param   slf4jLogger     the {@code Logger}
+     * @param   gradleLogger    the {@code Logger}
      * @param   text            the message
      */
     @Override
-    public void debug(Object slf4jLogger, String text) {
+    public void debug(Object gradleLogger, String text) {
         try {
-            debugMethod.invoke(slf4jLogger, text);
+            logMethod.invoke(gradleLogger, infoLevel, text);
         }
         catch (IllegalAccessException | InvocationTargetException e) {
             throw new Slf4jLoggerException(e);
@@ -223,13 +235,13 @@ public class Slf4jProxy implements Slf4jProxyInterface {
     /**
      * Output an info message.
      *
-     * @param   slf4jLogger     the {@code Logger}
+     * @param   gradleLogger    the {@code Logger}
      * @param   text            the message
      */
     @Override
-    public void info(Object slf4jLogger, String text) {
+    public void info(Object gradleLogger, String text) {
         try {
-            infoMethod.invoke(slf4jLogger, text);
+            logMethod.invoke(gradleLogger, lifecycleLevel, text);
         }
         catch (IllegalAccessException | InvocationTargetException e) {
             throw new Slf4jLoggerException(e);
@@ -239,13 +251,13 @@ public class Slf4jProxy implements Slf4jProxyInterface {
     /**
      * Output a warning message.
      *
-     * @param   slf4jLogger     the {@code Logger}
+     * @param   gradleLogger    the {@code Logger}
      * @param   text            the message
      */
     @Override
-    public void warn(Object slf4jLogger, String text) {
+    public void warn(Object gradleLogger, String text) {
         try {
-            warnMethod.invoke(slf4jLogger, text);
+            logMethod.invoke(gradleLogger, warnLevel, text);
         }
         catch (IllegalAccessException | InvocationTargetException e) {
             throw new Slf4jLoggerException(e);
@@ -255,13 +267,13 @@ public class Slf4jProxy implements Slf4jProxyInterface {
     /**
      * Output an error message.
      *
-     * @param   slf4jLogger     the {@code Logger}
+     * @param   gradleLogger    the {@code Logger}
      * @param   text            the message
      */
     @Override
-    public void error(Object slf4jLogger, String text) {
+    public void error(Object gradleLogger, String text) {
         try {
-            errorMethod.invoke(slf4jLogger, text);
+            logMethod.invoke(gradleLogger, errorLevel, text);
         }
         catch (IllegalAccessException | InvocationTargetException e) {
             throw new Slf4jLoggerException(e);
@@ -271,14 +283,14 @@ public class Slf4jProxy implements Slf4jProxyInterface {
     /**
      * Output an error message along with a {@link Throwable}.
      *
-     * @param   slf4jLogger     the {@code Logger}
+     * @param   gradleLogger    the {@code Logger}
      * @param   text            the message
      * @param   throwable       the {@link Throwable}
      */
     @Override
-    public void error(Object slf4jLogger, String text, Throwable throwable) {
+    public void error(Object gradleLogger, String text, Throwable throwable) {
         try {
-            errorThrowableMethod.invoke(slf4jLogger, text, throwable);
+            logThrowableMethod.invoke(gradleLogger, errorLevel, text, throwable);
         }
         catch (IllegalAccessException | InvocationTargetException e) {
             throw new Slf4jLoggerException(e);
@@ -288,14 +300,14 @@ public class Slf4jProxy implements Slf4jProxyInterface {
     /**
      * Output a message at the specified level.
      *
-     * @param   slf4jLogger     the {@code Logger}
+     * @param   gradleLogger    the {@code Logger}
      * @param   level           the {@code Level}
      * @param   text            the message
      */
     @Override
-    public void log (Object slf4jLogger, Level level, String text) {
+    public void log (Object gradleLogger, Level level, String text) {
         try {
-            dynamicLogMethods[level.ordinal()].invoke(slf4jLogger, text);
+            logMethod.invoke(gradleLogger, levels[level.ordinal()], text);
         }
         catch (IllegalAccessException | InvocationTargetException e) {
             throw new Slf4jLoggerException(e);